requirements.txt
# Echo Pipeline Testing Dependencies

# Core pipeline dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Content processing
markdown>=3.4.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0

# Testing and validation
jsonschema>=4.17.0
validators>=0.20.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Logging and reporting
loguru>=0.7.0
rich>=13.0.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0